chore #169: LikedsDTO, WriterDTO 파일 분리
