+++ iOS/traveline/Sources/Data/Network/DataMapping/WriterDTO.swift
@@ -1,0 +1,17 @@
+//
+//  WriterDTO.swift
+//  traveline
+//
+//  Created by 김태현 on 11/30/23.
+//  Copyright © 2023 traveline. All rights reserved.
+//
+
+import Foundation
+
+struct WriterDTO: Decodable {
+    let id: String
+    let name: String
+    let avatar: String?
+    let resourceId: String
+    let socialType: Int
+}
